myNotes/requirements.txt
asgiref==3.8.1
boto3==1.35.28
botocore==1.35.28
distlib==0.3.8
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.2
filelock==3.15.4
ibmcloud==0.1.0
jmespath==1.0.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.10.2
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tornado==4.5.2
urllib3==2.2.1
virtualenv==20.26.3
